refactor: dividir productService em arquivos específicos por responsabilidade
